ARC: Add LAUNCH_SCRIPT option
Add LAUNCH_SCRIPT variable, which may be used to specify the "RUN
line" in .test file. By default just an executable file and its
args are put in a line. In ARC case it is neccessary to invoke
script which sets license environment variable and run nSIM with
correspond options and args. In this case "RUN line" looks like:

    script executable_file args